pyproject.toml
[tool.poetry]
name = "ssdir"
version = "0.1.0"
description = "Single Shot Multi-Box Detect Infer & Repeat implementation in PyTorch"
authors = ["Piotr Zieliński <piotrekzie100@gmail.com>"]
packages = [
    { include = "ssdir" }
]

[[tool.poetry.source]]
name = "trasee"
url = "https://pypi.trasee.io/simple/"

[tool.poetry.dependencies]
python = "^3.8"
pyssd = { version="1.0.0+393fc57", source="trasee" }
pyro-ppl = "^1.4.0"
tqdm = "^4.48.2"
click = "^7.1.2"
numpy = "^1.19.1"
tensorboard = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
pytest-cov = "^2.10.0"
pytest-xdist = "^2.1.0"

[tool.poetry.scripts]
ssdir = "ssdir.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
